+++ src/lwc/geGiftEntryFormApp/geGiftEntryFormApp.js
@@ -5,10 +5,8 @@ import { LightningElement, api, track, wire } from 'lwc';
 import { getRecord, getFieldValue, updateRecord } from 'lightning/uiRecordApi';
 import { getObjectInfo } from 'lightning/uiObjectInfoApi';
 import { NavigationMixin } from 'lightning/navigation';
-import { fireEvent, registerListener, unregisterListener } from 'c/pubsubNoPageRef';
-import { HttpRequestError, CardChargedBDIError } from 'c/utilCustomErrors';
-import { isNotEmpty, validateJSONString, format, getNamespace } from 'c/utilCommon';
-import { getCurrencyLowestCommonDenominator } from 'c/utilNumberFormatter';
+import { registerListener, unregisterListener } from 'c/pubsubNoPageRef';
+import { validateJSONString, format, getNamespace } from 'c/utilCommon';
 import { handleError } from "c/utilTemplateBuilder";
 import GeLabelService from 'c/geLabelService';
 import geBatchGiftsHeader from '@salesforce/label/c.geBatchGiftsHeader';
@@ -16,10 +14,6 @@ import geBatchGiftsExpectedTotalsMessage
     from '@salesforce/label/c.geBatchGiftsExpectedTotalsMessage';
 import geBatchGiftsExpectedCountOrTotalMessage
     from '@salesforce/label/c.geBatchGiftsExpectedCountOrTotalMessage';
-import sendPurchaseRequest from '@salesforce/apex/GE_GiftEntryController.sendPurchaseRequest';
-import upsertDataImport from '@salesforce/apex/GE_GiftEntryController.upsertDataImport';
-import submitDataImportToBDI from '@salesforce/apex/GE_GiftEntryController.submitDataImportToBDI';
-import getPaymentTransactionStatusValues from '@salesforce/apex/GE_PaymentServices.getPaymentTransactionStatusValues';
 
 /*******************************************************************************
 * @description Schema imports
@@ -34,39 +28,10 @@ import EXPECTED_TOTAL_BATCH_AMOUNT
 import BATCH_ID_FIELD from '@salesforce/schema/DataImportBatch__c.Id';
 import BATCH_TABLE_COLUMNS_FIELD from '@salesforce/schema/DataImportBatch__c.Batch_Table_Columns__c';
 import REQUIRE_TOTAL_MATCH from '@salesforce/schema/DataImportBatch__c.RequireTotalMatch__c';
-import DI_PAYMENT_AUTHORIZE_TOKEN_FIELD from '@salesforce/schema/DataImport__c.Payment_Authorization_Token__c';
-import DI_PAYMENT_ELEVATE_ID from '@salesforce/schema/DataImport__c.Payment_Elevate_ID__c';
-import DI_PAYMENT_CARD_NETWORK from '@salesforce/schema/DataImport__c.Payment_Card_Network__c';
-import DI_PAYMENT_EXPIRATION_YEAR from '@salesforce/schema/DataImport__c.Payment_Card_Expiration_Year__c';
-import DI_PAYMENT_EXPIRATION_MONTH from '@salesforce/schema/DataImport__c.Payment_Card_Expiration_Month__c';
-import DI_PAYMENT_GATEWAY_ID from '@salesforce/schema/DataImport__c.Payment_Gateway_ID__c';
-import DI_PAYMENT_GATEWAY_TRANSACTION_ID from '@salesforce/schema/DataImport__c.Payment_Gateway_Payment_ID__c';
-import DI_PAYMENT_AUTHORIZED_AT from '@salesforce/schema/DataImport__c.Payment_Authorized_UTC_Timestamp__c';
-import DI_PAYMENT_LAST_4 from '@salesforce/schema/DataImport__c.Payment_Card_Last_4__c';
-import DI_PAYMENT_STATUS_FIELD from '@salesforce/schema/DataImport__c.Payment_Status__c';
-import DI_PAYMENT_DECLINED_REASON_FIELD from '@salesforce/schema/DataImport__c.Payment_Declined_Reason__c';
-import DI_PAYMENT_METHOD_FIELD from '@salesforce/schema/DataImport__c.Payment_Method__c';
-import DI_DONATION_AMOUNT_FIELD from '@salesforce/schema/DataImport__c.Donation_Amount__c';
-import DI_DONATION_CAMPAIGN_NAME_FIELD from '@salesforce/schema/DataImport__c.Donation_Campaign_Name__c';
 
 /*******************************************************************************
 * @description Constants
 */
-import { LABEL_NEW_LINE, HTTP_CODES } from 'c/geConstants';
-const PAYMENT_STATUS__C = DI_PAYMENT_STATUS_FIELD.fieldApiName;
-const PAYMENT_DECLINED_REASON__C = DI_PAYMENT_DECLINED_REASON_FIELD.fieldApiName;
-const PAYMENT_AUTHORIZE_TOKEN__C = DI_PAYMENT_AUTHORIZE_TOKEN_FIELD.fieldApiName;
-const PAYMENT_METHOD__C = DI_PAYMENT_METHOD_FIELD.fieldApiName;
-const PAYMENT_ELEVATE_ID = DI_PAYMENT_ELEVATE_ID.fieldApiName;
-const PAYMENT_CARD_NETWORK = DI_PAYMENT_CARD_NETWORK.fieldApiName;
-const PAYMENT_LAST_4 = DI_PAYMENT_LAST_4.fieldApiName;
-const PAYMENT_EXPIRATION_MONTH = DI_PAYMENT_EXPIRATION_MONTH.fieldApiName;
-const PAYMENT_EXPIRATION_YEAR = DI_PAYMENT_EXPIRATION_YEAR.fieldApiName;
-const PAYMENT_GATEWAY_ID = DI_PAYMENT_GATEWAY_ID.fieldApiName;
-const PAYMENT_TRANSACTION_ID = DI_PAYMENT_GATEWAY_TRANSACTION_ID.fieldApiName;
-const PAYMENT_AUTHORIZED_AT = DI_PAYMENT_AUTHORIZED_AT.fieldApiName;
-const DONATION_AMOUNT__C = DI_DONATION_AMOUNT_FIELD.fieldApiName;
-const DONATION_CAMPAIGN_NAME__C = DI_DONATION_CAMPAIGN_NAME_FIELD.fieldApiName;
 const BDI_DATA_IMPORT_PAGE = 'BDI_DataImport';
 
 export default class GeGiftEntryFormApp extends NavigationMixin(LightningElement) {
@@ -101,13 +66,6 @@ export default class GeGiftEntryFormApp extends NavigationMixin(LightningElement
         this.namespace = getNamespace(DATA_IMPORT_BATCH_OBJECT.objectApiName);
     }
 
-    connectedCallback() {
-        getPaymentTransactionStatusValues()
-            .then(response => {
-                this.PAYMENT_TRANSACTION_STATUS_ENUM = Object.freeze(JSON.parse(response));
-            });
-    }
-
     /*******************************************************************************
     * @description Receives a 'submit' event from geFormRenderer and proceeds down
     * the Batch or Single save paths depending on the current app mode.
@@ -124,7 +82,7 @@ export default class GeGiftEntryFormApp extends NavigationMixin(LightningElement
             if (this.isBatchMode) {
                 this.batchGiftSubmit(event);
             } else {
-                this.singleGiftSubmit(event);
+                // TODO: potentially receive event here and navigate to record detail page
             }
         } catch (error) {
             this.errorCallback(error);
@@ -142,365 +100,6 @@ export default class GeGiftEntryFormApp extends NavigationMixin(LightningElement
         const table = this.template.querySelector('c-ge-batch-gift-entry-table');
         this.dataImportRecord = event.detail.dataImportRecord;
         table.handleSubmit(event);
-    }
-
-    /*******************************************************************************
-    * @description Handles a single gift entry submit. Saves a Data Import record,
-    * makes an elevate payment if needed, and processes the Data Import through
-    * BDI.
-    *
-    * @param {object} event: Custom Event containing the Data Import record and a
-    * callback for handling and displaying errors in the form.
-    */
-    singleGiftSubmit = async (event) => {
-        let { inMemoryDataImport } = event.detail;
-        this.hasUserSelectedDonation = event.detail.hasUserSelectedDonation;
-        this._isCreditCardWidgetInDoNotChargeState = event.detail.isWidgetInDoNotChargeState;
-        try {
-            await this.saveDataImport(inMemoryDataImport);
-
-            const hasPaymentToProcess = this.dataImportRecord[PAYMENT_AUTHORIZE_TOKEN__C];
-            if (isNotEmpty(hasPaymentToProcess)) {
-                await this.processPayment();
-            }
-
-            if (!this.isFailedPurchase || this._isCreditCardWidgetInDoNotChargeState) {
-                await this.processDataImport();
-            }
-        } catch (error) {
-            this.errorCallback(error);
-        }
-    }
-
-    /*******************************************************************************
-    * @description Upserts the provided data import record. Attempts to retrieve
-    * an Elevate token for a purchase call if needed.
-    *
-    * @param {object} inMemoryDataImport: DataImport__c object built from the form
-    * fields.
-    *
-    * @return {object} dataImportRecord: A DataImport__c record
-    */
-    saveDataImport = async (inMemoryDataImport) => {
-        if (this.dataImportRecord.Id) {
-            this.loadingText = this.CUSTOM_LABELS.geTextUpdating;
-            inMemoryDataImport = this.prepareInMemoryDataImportForUpdate(inMemoryDataImport);
-        } else {
-            this.loadingText = this.CUSTOM_LABELS.geTextSaving;
-        }
-
-        this.dataImportRecord = await upsertDataImport({ dataImport: inMemoryDataImport });
-    };
-
-    /*******************************************************************************
-    * @description Re-apply Data Import id and relevant payment/elevate fields.
-    * The inMemoryDataImport is built new from the form on every save click. We
-    * need to catch it up with the correct id to make sure we update instead of
-    * inserting a new record on re-save attempts.
-    *
-    * @param {object} inMemoryDataImport: DataImport__c object built from the form
-    * fields.
-    *
-    * @return {object} inMemoryDataImport: DataImport__c object built from the form
-    * fields.
-    */
-    prepareInMemoryDataImportForUpdate(inMemoryDataImport) {
-        inMemoryDataImport.Id = this.dataImportRecord.Id;
-        inMemoryDataImport[PAYMENT_METHOD__C] = this.dataImportRecord[PAYMENT_METHOD__C];
-        inMemoryDataImport[PAYMENT_STATUS__C] =
-            this._isCreditCardWidgetInDoNotChargeState ? '' : this.dataImportRecord[PAYMENT_STATUS__C];
-        inMemoryDataImport[PAYMENT_DECLINED_REASON__C] =
-            this._isCreditCardWidgetInDoNotChargeState ? '' : this.dataImportRecord[PAYMENT_DECLINED_REASON__C];
-        inMemoryDataImport[PAYMENT_AUTHORIZE_TOKEN__C] =
-            this._isCreditCardWidgetInDoNotChargeState ? '' : this.dataImportRecord[PAYMENT_AUTHORIZE_TOKEN__C];
-        return inMemoryDataImport;
-    }
-
-    /*******************************************************************************
-    * @description Method attempts to make a purchase call to Payment
-    * Services. Immediately attempts to the charge the card provided in the Payment
-    * Services iframe (GE_TokenizeCard).
-    */
-    processPayment = async () => {
-        this.loadingText = this.CUSTOM_LABELS.geTextChargingCard;
-
-        const isReadyToCharge = await this.checkPaymentTransactionStatus(this.dataImportRecord[PAYMENT_STATUS__C]);
-        if (isReadyToCharge) {
-
-            const purchaseResponse = await this.makePurchaseCall();
-            if (purchaseResponse) {
-
-                let errors = this.processPurchaseResponse(purchaseResponse);
-
-                this.dataImportRecord = await upsertDataImport({ dataImport: this.dataImportRecord });
-
-                if (isNotEmpty(errors)) {
-                    this.isFailedPurchase = true;
-                    this.handleFailedPurchaseCall(purchaseResponse);
-                } else {
-                    this.isFailedPurchase = false;
-                }
-            }
-        }
-    }
-
-    /*******************************************************************************
-     * @description Updates the dataImportRecord fields with response values from
-     * payment services.
-     *
-     * @param {object} response The response object from payment services returned when
-     * purchase call is made.
-     *
-     * @return {string} A concatenated string of errors returned from the purchase call to
-     * payment services
-     */
-    processPurchaseResponse(response) {
-        let errors = '';
-        let responseBody = response.body;
-
-        this.dataImportRecord[PAYMENT_STATUS__C] = this.getPaymentStatus(response);
-        this.dataImportRecord[PAYMENT_ELEVATE_ID] = responseBody.id;
-
-        if (response.statusCode === HTTP_CODES.Created) {
-
-            if (isNotEmpty(responseBody.cardData)) {
-                this.dataImportRecord[PAYMENT_CARD_NETWORK] = responseBody.cardData.brand;
-                this.dataImportRecord[PAYMENT_LAST_4] = responseBody.cardData.last4;
-                this.dataImportRecord[PAYMENT_EXPIRATION_MONTH] = responseBody.cardData.expirationMonth;
-                this.dataImportRecord[PAYMENT_EXPIRATION_YEAR] = responseBody.cardData.expirationYear;
-            }
-            this.dataImportRecord[PAYMENT_DECLINED_REASON__C] = '';
-            this.dataImportRecord[PAYMENT_GATEWAY_ID] = responseBody.gatewayId;
-            this.dataImportRecord[PAYMENT_TRANSACTION_ID] = responseBody.gatewayTransactionId;
-            this.dataImportRecord[PAYMENT_AUTHORIZED_AT] = responseBody.authorizedAt;
-
-        } else {
-            this.dataImportRecord[PAYMENT_DECLINED_REASON__C] =
-                this.getPaymentDeclinedReason(response);
-
-            errors = this.getFailedPurchaseMessage(response);
-        }
-
-        return errors;
-    }
-
-    /*******************************************************************************
-    * @description Method attempts to handle a failed purchase call response.
-    *
-    * @param {object} purchaseResponse: Response from
-    * GE_GiftEntryController.sendPurchaseRequest()
-    */
-    handleFailedPurchaseCall(purchaseResponse) {
-        const hasPurchaseCallValidationErrors = purchaseResponse.statusCode === HTTP_CODES.Bad_Request;
-        if (hasPurchaseCallValidationErrors) {
-            this.catchPurchaseCallValidationErrors(purchaseResponse);
-        } else {
-            this.throwHttpRequestError(purchaseResponse);
-        }
-    }
-
-    /*******************************************************************************
-    * @description Dispatch an event to the form renderer whenever we have a field
-    * validation error from the purchase call. Various http 400 errors. Thus far
-    * the body of the response from these calls have had an errors array property.
-    *
-    * @param {object} purchaseResponse: Response from
-    * GE_GiftEntryController.sendPurchaseRequest()
-    */
-    catchPurchaseCallValidationErrors(purchaseResponse) {
-        let errors = this.getFailedPurchaseMessage(purchaseResponse);
-        let labelReplacements = [this.CUSTOM_LABELS.commonPaymentServices, errors];
-        let formattedErrorResponse = format(this.CUSTOM_LABELS.gePaymentProcessError, labelReplacements);
-
-        // We use the hex value for line feed (new line) 0x0A
-        let splitErrorResponse = formattedErrorResponse.split(LABEL_NEW_LINE);
-
-        const form = this.template.querySelector('c-ge-form-renderer');
-        form.showSpinner = false;
-        fireEvent(null, 'paymentError', {
-            error: {
-                message: splitErrorResponse,
-                isObject: true
-            }
-        });
-    }
-
-    /*******************************************************************************
-    * @description Throw an HttpRequestError for non-400 and non-200 responses from
-    * a purchase call.
-    *
-    * @param {object} purchaseResponse: Response from
-    * GE_GiftEntryController.sendPurchaseRequest()
-    */
-    throwHttpRequestError(purchaseResponse) {
-        const errorMessage =
-            this.CUSTOM_LABELS.commonPaymentServices + ': ' +
-            this.getFailedPurchaseMessage(purchaseResponse)
-
-        throw new HttpRequestError(
-            errorMessage,
-            purchaseResponse.status,
-            purchaseResponse.statusCode);
-    }
-
-    /*******************************************************************************
-    * @description Method checks the current payment transaction's status and
-    * returns true if the card is in a 'chargeable' status.
-    *
-    * @param {string} paymentStatus: Payment transaction status
-    *
-    * @return {boolean}: True if card is in a 'chargeable' status
-    */
-    checkPaymentTransactionStatus = async (paymentStatus) => {
-
-        switch (paymentStatus) {
-            case this.PAYMENT_TRANSACTION_STATUS_ENUM.PENDING: return true;
-            case this.PAYMENT_TRANSACTION_STATUS_ENUM.AUTHORIZED: return false;
-            case this.PAYMENT_TRANSACTION_STATUS_ENUM.CANCELED: return false;
-            case this.PAYMENT_TRANSACTION_STATUS_ENUM.CAPTURED: return false;
-            case this.PAYMENT_TRANSACTION_STATUS_ENUM.DECLINED: return true;
-            case this.PAYMENT_TRANSACTION_STATUS_ENUM.NONRETRYABLEERROR: return false;
-            case this.PAYMENT_TRANSACTION_STATUS_ENUM.RETRYABLEERROR: return true;
-            case this.PAYMENT_TRANSACTION_STATUS_ENUM.REFUNDISSUED: return false;
-            default: return true;
-        }
-    }
-
-    /*******************************************************************************
-    * @description Posts an http request through the `sendPurchaseRequest` apex
-    * method and parses the response.
-    *
-    * @return {object} response: An http response object
-    */
-    makePurchaseCall = async () => {
-        let purchaseResponseString = await sendPurchaseRequest({
-            requestBodyParameters: this.buildPurchaseRequestBodyParameters(),
-            dataImportRecordId: this.dataImportRecord.Id
-        });
-        let response = JSON.parse(purchaseResponseString);
-        if (response.body && validateJSONString(response.body)) {
-            response.body = JSON.parse(response.body);
-        }
-
-        return response;
-    }
-
-    /*******************************************************************************
-    * @description Builds parts of the purchase request body that requires data
-    * from the Data Import record upfront. We pass this into the `sendPurchaseRequest`
-    * method and is eventually merged in with the rest of the purchase request body.
-    *
-    * @return {object}: Object that we can deserialize and apply to the purchase
-    * request body in apex.
-    */
-    buildPurchaseRequestBodyParameters() {
-        const { firstName, lastName } = this.getCardholderNames();
-        const metadata = {
-            campaignCode: this.dataImportRecord[DONATION_CAMPAIGN_NAME__C]
-        };
-
-        return JSON.stringify({
-            amount: getCurrencyLowestCommonDenominator(this.dataImportRecord[DONATION_AMOUNT__C]),
-            firstName: firstName,
-            lastName: lastName,
-            metadata: metadata,
-            paymentMethodToken: this.dataImportRecord[PAYMENT_AUTHORIZE_TOKEN__C],
-        });
-    }
-
-    /*******************************************************************************
-    * @description Queries for the geFormRenderer component and retrieves the
-    * fabricated cardholder names object.
-    *
-    * @return {object}: Object containing firstName, lastName, and accountName
-    * from the form.
-    */
-    getCardholderNames() {
-        const renderer = this.template.querySelector('c-ge-form-renderer');
-        return renderer.getCardholderNames();
-    }
-
-    /*******************************************************************************
-    * @description Get the value for DataImport__c.Payment_Status__c from the
-    * purchase call response.
-    *
-    * @param {object} response: Http response object
-    *
-    * @return {string}: Status of the payment charge request
-    */
-    getPaymentStatus(response) {
-        return response.body.status || response.status || this.CUSTOM_LABELS.commonUnknownError;
-    }
-
-    /*******************************************************************************
-    * @description Get the value for DataImport__c.Payment_Declined_Reason__c from
-    * the purchase call response.
-    *
-    * @param {object} response: Http response object
-    *
-    * @return {string}: Reason the payment was declined
-    */
-    getPaymentDeclinedReason(response) {
-        const isSuccessfulPurchase = response.statusCode === HTTP_CODES.Created;
-        return isSuccessfulPurchase ? null : this.getFailedPurchaseMessage(response);
-    }
-
-    /*******************************************************************************
-    * @description Get the message or errors from a failed purchase call.
-    *
-    * @param {object} response: Http response object
-    *
-    * @return {string}: Message from a failed purchase call response
-    */
-    getFailedPurchaseMessage(response) {
-        // For some reason the key in the body object for 'Message'
-        // in the response we receive from Elevate is capitalized.
-        // Also checking for lowercase M in message in case they fix it.
-        return response.body.Message ||
-            response.body.message ||
-            response.errorMessage ||
-            JSON.stringify(response.body.errors.map(error => error.message)) ||
-            this.CUSTOM_LABELS.commonUnknownError;
-    }
-
-    /*******************************************************************************
-    * @description Sends the Data Import into BDI for processing and navigates to
-    * the opportunity record detail page on success.
-    */
-    processDataImport = async () => {
-        this.loadingText = this.CUSTOM_LABELS.geTextProcessing;
-
-        submitDataImportToBDI({ dataImport: this.dataImportRecord, updateGift: this.hasUserSelectedDonation })
-            .then(opportunityId => {
-                this.loadingText = this.CUSTOM_LABELS.geTextNavigateToOpportunity;
-                this.navigateToRecordPage(opportunityId);
-            })
-            .catch(error => {
-                this.catchBDIProcessingError(error);
-            });
-    }
-
-    /*******************************************************************************
-    * @description Catches a BDI processing error and transforms the error into
-    * something we can ingest in the form renderer.
-    */
-    catchBDIProcessingError(error) {
-        const hasPaymentBeenCaptured = this.checkForCapturedPayment();
-        if (hasPaymentBeenCaptured) {
-            error = new CardChargedBDIError(error);
-        }
-        this.errorCallback(error);
-    }
-
-    /*******************************************************************************
-    * @description Method checks to see if the current data import record has a
-    * successful payment transaction.
-    *
-    * TODO: Update to check for the elevate transaction id in the future.
-    */
-    checkForCapturedPayment() {
-        return this.dataImportRecord[PAYMENT_STATUS__C] === this.PAYMENT_TRANSACTION_STATUS_ENUM.CAPTURED &&
-            isNotEmpty(this.dataImportRecord[PAYMENT_AUTHORIZE_TOKEN__C]);
     }
 
     handleSectionsRetrieved(event) {
